+++ JsonToolsNppPlugin/Forms/SortForm.cs
@@ -23,8 +23,8 @@ namespace JSON_Tools.Forms
 
         private void SortButton_Clicked(object sender, EventArgs e)
         {
-            (bool fatal, JNode json, bool usesSelections, _) = Main.TryParseJson(preferPreviousDocumentType:true);
-            if (fatal || json == null)
+            (ParserState parserState, JNode json, bool usesSelections, _) = Main.TryParseJson(preferPreviousDocumentType:true);
+            if (parserState == ParserState.FATAL || json == null)
                 return;
             string pathQuery = "@" + PathTextBox.Text;
             JNode jsonAtPath;