+++ JsonToolsNppPlugin/Properties/AssemblyInfo.cs
@@ -28,5 +28,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("6.1.1.17")]
-[assembly: AssemblyFileVersion("6.1.1.17")]
+[assembly: AssemblyVersion("6.1.1.18")]
+[assembly: AssemblyFileVersion("6.1.1.18")]